x86: Move platforms to Kconfig.platforms
The x86 platform options in Kconfig are rather confusing, as they are a
mix of options that just control visibility of drivers and those that are
fundamental to the operation, with some of the code in arch/x86/platforms
and other code in drivers/platform/x86, and the options listed in various
places in arch/x86/Kconfig.

Make this a little more consistent by moving all options that control a
platform from arch/x86/Kconfig to arch/x86/Kconfig.platforms and put them
into a submenu.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>